// WADNR.Web/src/app/shared/generated/model/project-image-detail.ts
export class ProjectImageDetail { 
    ProjectImageID?: number;
    ProjectID?: number;
    FileResourceID?: number;
    FileResourceGuid?: string;
    Caption?: string | null;
    Credit?: string | null;
    IsKeyPhoto?: boolean;
    ExcludeFromFactSheet?: boolean;
    ProjectImageTimingID?: number | null;
    ProjectImageTimingDisplayName?: string | null;
    CreatedDate?: string;
    OriginalFilename?: string | null;
    ContentLength?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}